Put channel tests under channel not transport
